requirements.txt
streamlit
Pillow
torch
torchvision
torchcam
pandas
seaborn
scikit-learn


